replaced existing camera with native device camera
