+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pdfonline",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "dependencies": {
     "@material-ui/core": "^4.11.4",
@@ -14,7 +14,6 @@
     "react-cropper": "^2.1.7",
     "react-dom": "^17.0.2",
     "react-draggable": "^4.4.3",
-    "react-easy-crop": "^3.3.3",
     "react-redux": "^7.2.4",
     "react-scripts": "4.0.3",
     "redux": "^4.1.0",
